- add generating guideTextBoxType function - seperate Components, types - remove unused scss - import cssproperties from react
