+++ packages/input/GuideTextBox/src/GuideTextBox.stories.tsx
@@ -1,5 +1,6 @@
-import type { Meta, StoryObj } from '@storybook/react';
 import React, { useRef, useState } from 'react';
+
+import type { Meta, StoryObj } from '@storybook/react';
 
 import { GuideTextBoxForGeneral } from './GuideTextBoxForGeneral';
 import { GuideTextBoxForPairedVerification } from './GuideTextBoxForPairedVerification';
