+++ packages/input/GuideTextBox/src/Type.ts
@@ -1,0 +1,8 @@
+
+export type guideTextType = 'initial' | 'normal' | 'success' | 'fail' | 'required';
+
+export type purposeType = 'uniqueness' | 'verification' | 'send';
+
+export type validationStatusType = 'undone' | 'pending' | 'success' | 'fail';
+
+export type buttonStatusType = 'activated' | 'inactivated' | 'pending' | 'success' | 'fail';
