+++ packages/input/GuideTextBox/src/main.ts
@@ -4,6 +4,6 @@ export { GuideTextBoxForPassword } from './GuideTextBoxForPassword';
 export { GuideTextBoxForPairedVerification } from './GuideTextBoxForPairedVerification';
 
 export { TextBox } from './GuideTextBox';
-export { VerificationButton } from './GuideTextBox'
+export { VerificationButton } from './VerificationButton';
 
-export type { guideTextType, purposeType, validationStatusType, buttonStatusType } from './GuideTextBox'
+export type { guideTextType, purposeType, validationStatusType, buttonStatusType } from './Type'
